Move weight parsing logic into an utility class
